tests/wpt/css-tests/cssom-1_dev/xhtml1print/index-001.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
  <title>CSS OM: CSS Values</title>
  <link href="mailto:manian@adobe.com" rel="author" title="Divya Manian" />
  <link href="http://www.w3.org/TR/cssom/#css-values" rel="help" />
  <meta content="dom" name="flags" />
  <meta content="The style value should be serialized to margin: 20px;" name="assert" />
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>         
<body>
    <div id="cssomtestElm"></div>
    <div id="log"></div>    
    <script>
        var testElm = document.getElementById('cssomtestElm');
        // Set the transform 
        document.getElementById('cssomtestElm').style.margin = "20px 20px 20px 20px";

        // Verify that the transform was set as expected
        test(function() {assert_equals(
            document.getElementById('cssomtestElm').style.cssText, //Actual
            "margin: 20px;",  //Expected
            "Margin should be serialized as 'margin: 20px;'")}, //Description
            "margin_20px_20px"); //name
    </script>
 
 

</body></html>
